+++ src/pages/api/unit/property.ts
@@ -16,11 +16,18 @@ export default async function handler(
   res: NextApiResponse
 ) {
 
-  const {
-    query: { id, searchQuery },
-  }: any = req;
+  // const {
+  //   query: { id, searchQuery },
+  // }: any = req;
 
   const decodedToken = authenticateUser(req, res);
+
+  if (!decodedToken?.user?._id) {
+    return res.status(401).json({
+      success: false,
+      msg: "Not Authorized",
+    });
+  }
 
   try {
     await mongoose
@@ -29,12 +36,35 @@ export default async function handler(
         "mongodb://localhost:27017/test_db"
       )
       .then(() => {
-        // USER
-        const { _id, role } = decodedToken.user;
 
+        //type of request
         const { method } = req;
+
+        // Property ID
+        const property = req.body.property || req.query.id
+
+        // Reject if no property provided
+        if (!property) {
+          return res.status(400).json({
+            success: false,
+            msg: "No property provided",
+          });
+        }
+
+        // Get permissions
+        const userRoles = decodedToken.userRoles
+        const userPropertyRoles = userRoles?.find((role: any) => role.property === property)
+        const permissions = userPropertyRoles?.role?.permissions
+
         switch (method) {
           case "GET":
+            const getPermission = permissions?.find((permission: any) => permission.name === "View unit")
+            if (!getPermission) {
+              return res.status(401).json({
+                success: false,
+                msg: "Not Authorized",
+              });
+            }
             fetchAllPropertyUnits(req, res);
             // fetchAllUnits(req, res)
             break;
@@ -42,12 +72,33 @@ export default async function handler(
             //   fetchSingleUnit(req, res);
             break;
           case "POST":
+            const postPermission = permissions?.find((permission: any) => permission.name === "Create unit")
+            if (!postPermission) {
+              return res.status(401).json({
+                success: false,
+                msg: "Not Authorized",
+              });
+            }
             createUnit(req, res);
             break;
           case "PUT":
+            const putPermission = permissions?.find((permission: any) => permission.name === "Edit unit")
+            if (!putPermission) {
+              return res.status(401).json({
+                success: false,
+                msg: "Not Authorized",
+              });
+            }
             updateUnit(req, res);
             break;
           case "DELETE":
+            const deletePermission = permissions?.find((permission: any) => permission.name === "Delete unit")
+            if (!deletePermission) {
+              return res.status(401).json({
+                success: false,
+                msg: "Not Authorized",
+              });
+            }
             deleteUnit(req, res);
             break;
           default:
